rpi/requirements.txt
paho-mqtt==1.5.1
pycomm3==1.1.1